Agrego logica de transporte

+++ src/model/Paquete.scala
@@ -1,5 +1,5 @@
 package model
 
-class Paquete(val volumen: Double) {
+class Paquete(val volumen: Int) {
 
 }
